Refactor NodePartitioningManager
Extract connector partitioning provider management code into a
separate class, PartitioningProviderManager
